updates
screen timeout 10 min
raise z when pause then lower back and add offset to compensate backlash of z axis
compensate z backlash on leveling
update feed/retract
remove z from backlash default for the moment because the backlash is ok just for round objects ,  the correction is applied to the whole next move, which can be very long, and the object gets distorted also.
reduce junctoin dev because of too fast moves leads to layer shift
set z to -1,99 from 0,100 to compensate backlash and a little offset to move the stepper after that the phisical axis is actually moving. (play in z axis)
to find a backlash fix to apply punctually and inject move in the stepper in the opposite direction just as backlash distance.